typing_requirements.txt
ruff
mypy
django-stubs
